1. DetailActivity.kt fragment로 변경
