+++ android/app/src/main/AndroidManifest.xml
@@ -53,9 +53,6 @@
             android:name=".LocationPermissionActivity"
             android:exported="true" />
 
-        <activity
-            android:name=".productSearch.DetailActivity"
-            android:exported="false" />
     </application>
 
 </manifest>
