+++ android/app/src/main/res/layout/fragment_detail.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".productSearch.DetailActivity">
+    tools:context=".productSearch.DetailFragment">
 
     <LinearLayout
         android:id="@+id/toolbar"
